test(cli): isolate tools command suite from the cli root import
